requirements.txt
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
openai>=1.50.0
torch>=2.0.0
transformers>=4.40.0
sentence-transformers>=2.2.2
